# allow `ls` to work a little faster via threads