@fdncred 2023-05-03
I added sort to make it more deterministic. We could easily add a config option for default sort, which has been asked for a few times already.
![image](https://user-images.githubusercontent.com/343840/236028493-2c492a3b-7b9e-4afc-a160-26f9175364fc.png)