@fdncred 2023-05-04
The hold up to landing this is that it's not streaming anymore. We may need something more on the level of the `jwalk` crate that supposedly threads, sorts, and streams all at once. But then the problem is that `jwalk` is a directory traversal crate versus a globbing crate. I'm not sure what the solution is.